Move resources to testclasses in fieldaccesses
