+++ src/test/java/spoon/test/fieldaccesses/testclasses/Mouse.java
@@ -1,4 +1,4 @@
-package spoon.test.fieldaccesses;
+package spoon.test.fieldaccesses.testclasses;
 
 public class Mouse {
